Restructure table component files
